requirements.txt
flask
yfinance
flask_cors
python-dotenv
mysql-connector-python
pymysql
sqlalchemy
flask_sqlalchemy